+++ CHANGES.md
@@ -9,6 +9,7 @@ Fixed Issues:
 * [#2169](https://github.com/ckeditor/ckeditor-dev/issues/2169): [Edge] Fixed: Error thrown when pasting into editor.
 * [#2107](https://github.com/ckeditor/ckeditor-dev/issues/2107): Fixed: [Autocomplete](https://ckeditor.com/cke4/addon/autocomplete): holding and releasing mouse button is not inserting autocomplete item.
 * [#2167](https://github.com/ckeditor/ckeditor-dev/issues/2167): Fixed: Matching in [Emoji](https://ckeditor.com/cke4/addon/emoji) plugin is not case insensitive.
+* [#1887](https://github.com/ckeditor/ckeditor-dev/issues/1887): Fixed: Incorrect selection styling in readonly editor for selection made by [Table Selection](https://ckeditor.com/cke4/addon/tableselection) plugin.
 
 ## CKEditor 4.10